requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
drf-yasg==1.21.8
gunicorn==23.0.0
inflection==0.5.1
packaging==24.2
PyMySQL==1.1.1
pytz==2024.2
PyYAML==6.0.2
railway==0.0.4
sqlparse==0.5.2
uritemplate==4.1.1
whitenoise==6.8.2
